ariketak/ezagutza/requirements.txt
dnspython
geopy
Pillow
shodan